+++ src/graphics/render/Mesh.h
@@ -1,0 +1,37 @@
+#pragma once
+#include <glm/glm.hpp>
+#include <string>
+#include <vector>
+#include "Shader.h"
+
+struct Vertex {
+	glm::vec3 pos;
+	glm::vec3 norm;
+	glm::vec2 texCoords;
+};
+
+struct Texture {
+	unsigned int id;
+	std::string type;
+	std::string path;
+};
+
+class Mesh {
+public:
+	/*  Mesh Data  */
+	std::vector<Vertex> vertices;
+	std::vector<unsigned int> indices;
+	std::vector<Texture> textures;
+
+	/*  Functions  */
+	Mesh(std::vector<Vertex> vertices, std::vector<unsigned int> indices, std::vector<Texture> textures);
+
+
+	void draw(Shader shader);
+private:
+	/*  Render data  */
+	unsigned int VAO, VBO, EBO;
+
+	/*  Functions    */
+	void setupMesh();
+};
