+++ src/objects/GameObject.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include <graphics/_obj/Shader.h>
-#include <graphics/Model.cpp>
+#include <graphics/render/Shader.h>
+#include <graphics/render/Model.h>
 #include <glm/glm.hpp>
 #include <glm/gtc/matrix_transform.hpp>
 #include <graphics/_options/graphics_vars.h>
